refactor: rename initializeFirebase to firebase-client
