+++ frontend/src/integrations/firebase/firebase-client.ts
